Refactor: 날짜 필터를 YEAR/MONTH 함수에서 범위 비교로 변경해 인덱스 활용

+++ core-api/src/main/java/com/pocketstock/core/budget/dto/BudgetGoalItem.java
@@ -1,3 +1,5 @@
 package com.pocketstock.core.budget.dto;
 
-public record BudgetGoalItem(String category, Long budget) {}
+import java.math.BigDecimal;
+
+public record BudgetGoalItem(String category, BigDecimal budget) {}
